Rearranged the ng2 input and form bootstrappers.
The input bootstrapper should do most of the input testing, while the form bootstrapper should focus on form behaviors, such as how forms work, how validation works, and how they work together. We'll probably want to do a better job at grouping the input examples, and/or eventually give each its own page.
