+++ bootstrapper/forms/formsNg2.html
@@ -1,79 +1,29 @@
 <h3>Forms</h3>
 <h4><a href="https://github.com/SamGraber/TypeScript-Angular-Components/tree/master/source/components/form/form.md">rlForm</a>:</h4>
+<h5>Form with validation</h5>
+<rlForm [save]="waitCallback">
+	<rlTextbox rlRequired="This is a required field" label="Required textbox" maxlength="10"></rlTextbox>
+	<rlValidationGroup [model]="rating" [validator]="validator">
+		<rlUserRating [(value)]="rating" [range]="3"></rlUserRating>
+	</rlValidationGroup>
+	<rlButtonSubmit>Submit</rlButtonSubmit>
+</rlForm>
 <h5>Form with two-way bindings</h5>
+<rlForm [save]="waitCallback">
+	<rlTextbox [(value)]="text" label="Textbox with 2-way binding"></rlTextbox>
+	<rlButtonSubmit>Submit</rlButtonSubmit>
+</rlForm>
+<h5>Nested forms</h5>
 <rlForm #testForm [save]="saveTestForm">
+	<rlTextbox name="textbox1" label="Textbox in outer form"></rlTextbox>
 	<rlForm>
-		<rlCheckbox [(value)]="checked" label="Checkbox with 2-way binding"></rlCheckbox>
-		<rlRadioGroup [(value)]="selection" label="Spinner">
-			<rlRadio label="Option 1" [option]="options[0]"></rlRadio>
-			<rlRadio label="Option 2" [option]="options[1]"></rlRadio>
-			<rlRadio label="Option 3" [option]="options[2]"></rlRadio>
-		</rlRadioGroup>
-		<rlTextbox [(value)]="text" label="Textbox with 2-way binding"></rlTextbox>
-		<rlTextbox rlRequired="This is a required field" label="Textbox" maxlength="10"></rlTextbox>
-		<rlTextarea [(value)]="text" label="Textarea" rlRequired="This is a required field" rows="3" maxlength="10"></rlTextarea>
-		<rlSpinner [(value)]="number" label="Spinner" roundToStep="true" step="3.65" decimals="2" prefix="$" postfix="/hr"></rlSpinner>
-		<rlSelect [(value)]="selection" label="Select" [options]="options" transform="value"></rlSelect>
-		<rlSelect [(value)]="selection" label="Select" [options]="options" transform="value">
-			<template let-option>#{{option.value}}</template>
-		</rlSelect>
-		<rlTypeahead [(value)]="typeaheadSelection" prefix="Search for a" label="Typeahead collapsable" [getItems]="searchOptions" transform="value" allowCollapse="true"></rlTypeahead>
-		<rlTypeahead (select)="log($event)" prefix="Search for a" label="Typeahead client searching" [getItems]="getOptions" clientSearch="true" transform="value">
-			<template let-option>Value - {{option.value}}</template>
-		</rlTypeahead>
-		<rlTypeahead [(value)]="typeaheadSelection" prefix="Search for a" label="Typeahead with create" [getItems]="getOptions" clientSearch="true" transform="value" [create]="createOption" allowCollapse="true"></rlTypeahead>
-		<rlTypeaheadList [(value)]="selections" label="Typeahead list" [getItems]="getOptions" clientSearch="true" transform="value"></rlTypeaheadList>
-		<rlTypeaheadList [(value)]="selections" label="Typeahead list with templates" [getItems]="getOptions" clientSearch="true" transform="value">
-			<div *rlListHeader class="col-xs-12">Other name</div>
-			<div *rlListItem="let option; let remove = remove">
-				<div class="col-xs-10">Value - {{option.value}}</div>
-				<div class="col-xs-2"><rlButtonAsync [action]="remove">Remove</rlButtonAsync></div>
-			</div>
-		</rlTypeaheadList>
-		<rlDateTime [(value)]="date" label="DateTime"></rlDateTime>
-		<rlAbsoluteTime [(value)]="time" label="AbsoluteTime"></rlAbsoluteTime>
-		<rlValidationGroup [model]="rating" [validator]="validator">
-			<rlUserRating [(value)]="rating" [range]="3"></rlUserRating>
-		</rlValidationGroup>
+		<rlTextbox name="textbox2" label="Textbox in nested form"></rlTextbox>
 	</rlForm>
 	<rlButtonSubmit>Submit</rlButtonSubmit>
-	{{checked}}
-	{{text}}
 </rlForm>
 <h5>Form with names</h5>
 <rlForm [save]="waitCallback">
-	<rlCheckbox name="checkbox1" label="Checkbox 1"></rlCheckbox>
-	<rlRadioGroup name="radio1">
-		<rlRadio label="Option 1" [option]="options[0]"></rlRadio>
-		<rlRadio label="Option 2" [option]="options[1]"></rlRadio>
-		<rlRadio label="Option 3" [option]="options[2]"></rlRadio>
-	</rlRadioGroup>
 	<rlTextbox name="textbox1" label="Textbox 1" rlRequired="Required textbox"></rlTextbox>
-	<rlTextarea name="textarea1" label="Textarea 1" rlRequired="Required text area"></rlTextarea>
-	<rlSpinner name="spinner1" label="Spinner 1"></rlSpinner>
-	<rlSelect name="select1" label="Select" [options]="optionsAsync" transform="value" rlRequired="Required select" nullOption="None"></rlSelect>
-	<rlTypeahead name="typeahead1" prefix="Search for a" label="Typeahead" [getItems]="searchOptions" transform="value" allowCollapse="true"></rlTypeahead>
-	<rlDateTime name="dateTime1" label="DateTime"></rlDateTime>
-	<rlAbsoluteTime name="time1" label="AbsoluteTime"></rlAbsoluteTime>
-	<rlUserRating name="userRating1"></rlUserRating>
-	<rlButtonSubmit rightAligned="true">Submit right</rlButtonSubmit>
-</rlForm>
-<h5>Disabled form</h5>
-<rlForm [save]="waitCallback">
-	<rlCheckbox name="checkbox1" disabled="true" label="Checkbox 1"></rlCheckbox>
-	<rlRadioGroup name="radio1" disabled="true">
-		<rlRadio label="Option 1" [option]="options[0]"></rlRadio>
-		<rlRadio label="Option 2" [option]="options[1]"></rlRadio>
-		<rlRadio label="Option 3" [option]="options[2]"></rlRadio>
-	</rlRadioGroup>
-	<rlTextbox name="textbox1" disabled="true" label="Textbox 1" rlRequired="This is a required field"></rlTextbox>
-	<rlTextarea name="textarea1" disabled="true" label="Textarea 1" rlRequired="This is a required field"></rlTextarea>
-	<rlSpinner name="spinner1" label="Spinner 1" disabled="true" prefix="$" postfix="/hr"></rlSpinner>
-	<rlSelect name="select1" label="Select" [options]="optionsAsync" transform="value" disabled="true"></rlSelect>
-	<rlTypeahead name="typeahead1" prefix="Search for a" label="Typeahead" [getItems]="searchOptions" transform="value" allowCollapse="true" disabled="true"></rlTypeahead>
-	<rlDateTime name="dateTime1" label="DateTime" disabled="true"></rlDateTime>
-	<rlAbsoluteTime name="time1" label="AbsoluteTime" disabled="true"></rlAbsoluteTime>
-	<rlUserRating name="userRating1" disabled="true"></rlUserRating>
 	<rlButtonSubmit rightAligned="true">Submit right</rlButtonSubmit>
 </rlForm>
 <h5>Broken validation</h5>
